requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.2.0
braceexpand==0.1.7
certifi==2024.2.2
click==8.1.7
discord.py==2.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.7
import_expression==1.1.5
jishaku==2.5.2
multidict==6.0.5
pymongo==4.7.2
python-dotenv==1.0.1
six==1.16.0
wheel==0.43.0
yarl==1.9.4
